backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
langchain==0.1.12
langchain-community==0.0.28
langchain-huggingface==0.0.2
langchain-text-splitters==0.0.1
pymongo==4.10.0
transformers>=4.40.0
sentence-transformers>=2.6.0
accelerate==0.30.1
torch==2.2.1
chromadb==0.4.22
python-dotenv==1.0.1
bs4==0.0.1
requests==2.31.0
motor==3.7.0
